menu & submenu api updated

+++ src/main/java/com/example/Test/Series/repositories/MenuRepository.java
@@ -4,6 +4,7 @@ import com.example.Test.Series.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
 }
