+++ src/main/java/com/example/Test/Series/repositories/SubmenuRepository.java
@@ -4,6 +4,7 @@ import com.example.Test.Series.entity.Submenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface SubmenuRepository extends JpaRepository<Submenu, Long> {
+public interface SubmenuRepository extends JpaRepository<Submenu, Integer> {
 }
