+++ src/main/java/com/example/Test/Series/services/SubmenuServices.java
@@ -2,13 +2,12 @@ package com.example.Test.Series.services;
 
 import com.example.Test.Series.entity.Submenu;
 import com.example.Test.Series.exceptions.MenuException;
-
 import java.util.List;
 
 public interface SubmenuServices {
     List<Submenu> getAllSubmenus() throws MenuException;
-    Submenu getSubmenuById(Long id) throws MenuException;
+    Submenu getSubmenuById(Integer id) throws MenuException;
     Submenu createSubmenu(Submenu submenu) throws MenuException;
-    Submenu updateSubmenu(Long id, Submenu submenu) throws MenuException;
-    void deleteSubmenu(Long id) throws MenuException;
+    Submenu updateSubmenu(Integer id, Submenu submenu) throws MenuException;
+    void deleteSubmenu(Integer id) throws MenuException;
 }
